chore(repo): modernize github front matter
GitHub has a bunch of extra metadata you can attach to the repository.
This is all that metadata, including helpful guides for filing bug
reports, contributing, etc. It also adds a `CODEOWNERS` file that can be
used to specify owners of specific parts of the repository, which is
likely useful in the long run -- and can be used to mandate reviews for
PRs even with single code owners, too.

Signed-off-by: Austin Seipp <aseipp@pobox.com>