refactor(comp): drop stp support
As noted by Julie in issue #16, STP is a historical default for the
compiler to solve constraints, but recent Yices is now in use as well
with the open source release (Yices is GPLv3). Since CUDD is also being
removed, we can take this chance to just drop STP as well.

Signed-off-by: Austin Seipp <aseipp@pobox.com>